ai_engine/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi
uvicorn[standard]
langchain
langchain-community
langchain-openai
langchain-google-genai
langgraph
langchain-core
sentence-transformers
accelerate
pgvector
rouge-score
psycopg2-binary
redis
pydantic
python-dotenv
httpx
structlog
sentry-sdk
reportlab
python-pptx
python-docx
isodate
python-dateutil
torch
Pillow
